+++ components/widget/screen/src/main/java/com/flipperdevices/widget/screen/viewmodel/WidgetSelectViewModel.kt
@@ -78,6 +78,7 @@ class WidgetSelectViewModel @VMInject constructor(
         info { "#onSelectKey for $widgetId $keyPath" }
         viewModelScope.launch {
             widgetDataApi.updateKeyForWidget(widgetId, keyPath)
+            widgetApi.resetStateOfWidget(widgetId)
             widgetApi.invalidate()
             notifyActivityAboutResult(currentActivity)