+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [Feature] New report bug screen
 - [Feature] Self update(Google play)
 - [Feature] Add SAVE_DUMP, MFKEY32, OPEN_NFC_DUMP_EDITOR
+- [Feature] Add report changes to synchronization end event
 - [GitHub] Update github repo picture
 - [CI] Migrate to Detekt
 - [CI] Android lint in CI
@@ -37,6 +38,7 @@
 - [FIX] Fix mfkey32
 - [FIX] Use CurrentActivityHolder for get activity in updater(replace Context cast)
 - [FIX] Move apps from experimental settings to debug
+- [FIX] Fix customize button when set widget state
 - [REFACTOR] Migrate bottom bar to compose navigation
 - [REFACTOR] Bump Android Gradle Plugin
 - [REFACTOR] Fix detekt compose issues